requirements.txt
tensorflow
keras-tuner
numpy
matplotlib
scikit-learn
librosa
soundfile
ffmpeg-python
pydub
streamlit
google-cloud-storage
apache-airflow-providers-google
apache-airflow-providers-cncf-kubernetes
apache-airflow-providers-docker
